+++ server/src/functions/utils/pdf.js
@@ -1,0 +1,6 @@
+const fs = require("fs")
+
+module.exports = {
+  convertFileToBase64: async (filePath) =>
+    await fs.readFileSync(filePath).toString("base64"),
+}